% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{aggregateStdRes}
\alias{aggregateStdRes}
\title{aggregateStdRes}
\usage{
aggregateStdRes(sample.id, range, gene, sample.stdres, all.stdres)
}
\arguments{
\item{all.stdres}{}
}
\description{
aggregateStdRes
}
