% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cleanSeg}
\alias{cleanSeg}
\title{cleanSeg}
\usage{
cleanSeg(seg.x, min.size = 50000, loss.thresh = -0.1,
  gain.thresh = 0.1)
}
\arguments{
\item{seg.x}{A seg file in data frame, will be converted ot GRanges}

\item{min.size}{The minimum size of a segment to be allowed}

\item{loss.thresh}{Anything lower than this will be classified as a loss}

\item{gain.thresh}{Anything higher than this will be classified as a gain}
}
\description{
cleanSeg
}
