% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.bar.R
\name{color.bar}
\alias{color.bar}
\title{color.bar}
\usage{
color.bar(lut, min, max = -min, nticks = 11, ticks = seq(min, max,
  len = nticks), title = "")
}
\arguments{
\item{lut}{Vector of colours}

\item{min}{Min value to plot}

\item{max}{Max value to plot}

\item{nticks}{numberof ticks}

\item{title}{}
}
\description{
color.bar
}
