% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compTwoGenes.R
\name{compTwoGenes}
\alias{compTwoGenes}
\title{compTwoGenes}
\usage{
compTwoGenes(gene.x, gene.y, col.df, cex.type = "xy", use.seg = FALSE,
  plot.legend = TRUE, use.absolute = FALSE)
}
\arguments{
\item{cex.type}{}
}
\description{
compTwoGenes
}
