% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAndPlotFeatures.R
\name{.getStdResDf}
\alias{.getStdResDf}
\title{.getStdResDf}
\usage{
.getStdResDf(chr, bin.left, bin.right, stdres)
}
\arguments{
\item{chr}{Chr ID}

\item{bin.left}{Left loci of each genomic bin}

\item{bin.right}{Right loci of each genomic bin}

\item{stdres}{The raw Standardized residuals of each genomic bin for AB}
}
\value{

}
\description{
.getStdResDf
}
