% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAndPlotFeatures.R
\name{.intersectLohWithSeg}
\alias{.intersectLohWithSeg}
\title{.intersectLohWithSeg}
\usage{
.intersectLohWithSeg(i, chr, bin.left, bin.right, seg, stdres)
}
\arguments{
\item{stdres}{}
}
\description{
.intersectLohWithSeg
}
