% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compTwoGenes.R
\name{.simplifyMutInfo}
\alias{.simplifyMutInfo}
\title{.simplifyMutInfo}
\usage{
.simplifyMutInfo(mut.attr, mut.x, mut.y, use.absolute = FALSE)
}
\arguments{
\item{mut.y}{}
}
\value{

}
\description{
.simplifyMutInfo
}
