% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{doTheChi}
\alias{doTheChi}
\title{doTheChi}
\usage{
doTheChi(ctbl = NULL, attributes, mut.attr,
  acceptable.terms = c("HOMDEL", "HETLOSS", "LOH", "Het",
  "no_alteration", "Unknown"), tbl.idx = c(1, 3), gene = "ADAM10",
  mut = "ADAM10_CNA")
}
\arguments{
\item{mut}{}
}
\description{
doTheChi
}
