% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{filterOverlaps}
\alias{filterOverlaps}
\title{filterOverlaps}
\usage{
filterOverlaps(i)
}
\arguments{
\item{i}{The returned object for GRanges overlap}
}
\value{
A filtered version of GRanges overlap with a
unique queryHits
}
\description{
filterOverlaps
}
