% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTcgaMuts.R
\name{filterTcgaMuts}
\alias{filterTcgaMuts}
\title{filterTcgaMuts}
\usage{
filterTcgaMuts(muts)
}
\arguments{
\item{muts}{}
}
\description{
filterTcgaMuts
}
\examples{
muts <- read.table("~/git/SchramekLOH/SchramekLOH/data-raw/data_mutations_extended.trimmed.2019-06-11.txt",
                   quote="", fill=FALSE, stringsAsFactors = FALSE, check.names = FALSE,
                   sep="\\t", header=TRUE)
x <- filterTcgaMuts(muts)
save(x[['muts']], x[['simple.mut']], 
     file="~/git/SchramekLOH/SchramekLOH/data/muts.extended.RData")
}
