% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{generateDatasets}
\alias{generateDatasets}
\title{generateDatasets}
\usage{
generateDatasets(git = "~/git/SchramekLOH/SchramekLOH/data")
}
\arguments{
\item{git}{Path to the git/SchramekLOH/data directory}
}
\description{
generateDatasets
}
