% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{generateIgvAttributes}
\alias{generateIgvAttributes}
\title{generateIgvAttributes}
\usage{
generateIgvAttributes(idx, sample.stdres, mapping.cov,
  stdres.thresh = -5)
}
\arguments{
\item{stdres.thresh}{}
}
\description{
generateIgvAttributes
}
