% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohLogistic.R
\name{getLohProb}
\alias{getLohProb}
\title{getLohProb}
\usage{
getLohProb(s, mod, pos = c("seqnames", "start", "end", "width"))
}
\arguments{
\item{s}{sample specific dataframe, split from combineSegLohPurity()}

\item{mod}{logistic regression model containing predictors: 'mean', 'purity'}

\item{pos}{}
}
\description{
getLohProb
}
