% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneLoci.R
\name{getMapping}
\alias{getMapping}
\title{getMapping}
\usage{
getMapping(in.col = "ENTREZID", out.cols = c("SYMBOL", "ENSEMBL"))
}
\arguments{
\item{in.col}{Column name for org.Hs.eg.db (e.g. ENTREZID)}

\item{out.cols}{Column name(s) for org.Hs.eg.db (e.g. SYMBOL)}
}
\description{
getMapping
}
