% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{loadBirdseed}
\alias{loadBirdseed}
\title{loadBirdseed}
\usage{
loadBirdseed(git = "~/git/SchramekLOH/SchramekLOH/data")
}
\value{

}
\description{
This is controlled access, so the resulting Rdata file will
  not be committed and psuehd in the git repo
}
