% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDatasets.R
\name{loadGeneExpr}
\alias{loadGeneExpr}
\title{loadGeneExpr}
\usage{
loadGeneExpr(git = "~/git/SchramekLOH/SchramekLOH/data")
}
\value{

}
\description{
This is open access, so the resulting Rdata file is pushed
}
