% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAndPlotFeatures.R
\name{mapAndPlotFeatures}
\alias{mapAndPlotFeatures}
\title{mapAndPlotFeatures}
\usage{
mapAndPlotFeatures(id, mapping.cov, use.affy, plotsdir, snp6.chr, seg.chr,
  bs.chr, z.chr, r, p, cn, gen.plot = TRUE, use.absolute = FALSE)
}
\arguments{
\item{gen.plot}{}
}
\description{
mapAndPlotFeatures
}
