% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{mapIds}
\alias{mapIds}
\title{mapIds}
\usage{
mapIds(id, mapping, in.type = "tcga", out.type = "affy")
}
\arguments{
\item{id}{input ID (e.g. TCGA-CN-6011-01)}

\item{mapping}{Mapping data frame}

\item{in.type}{either 'tcga' (TCGA-CN-6011-01) or
affy ('FISTS_p_TCGA_b107_121_SNP_N_GenomeWideSNP_6_D08_777884')}

\item{out.type}{either 'tcga' or 'affy'}
}
\description{
mapIds
}
\examples{
mapIds(id, mapping.cov, in.type='tcga', out.type='affy')
}
