% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{parseIdsByMutation}
\alias{parseIdsByMutation}
\title{parseIdsByMutation}
\usage{
parseIdsByMutation(gene, gene.ex, seg.ids = seg.ids, lo.q = 0.1,
  hi.q = 0.9, cbio = NULL, keep.only.abs = FALSE)
}
\arguments{
\item{hi.q}{}
}
\description{
parseIdsByMutation
}
