% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMETable.R
\name{plotMEMatrix}
\alias{plotMEMatrix}
\title{Plots Mutual exclusivity heatmaps}
\usage{
plotMEMatrix(mat, min = -0.5, max = 0.5, step = 0.01,
  lo = "#2c7bb6", mid = "white", hi = "#1a9641")
}
\arguments{
\item{mat}{The matrix from genMEMatrix, or the table from table.tsv}

\item{min}{Min log2 odds ratio to plot}

\item{max}{Max log2 odds ratio}

\item{step}{Increment in steps from min to max}

\item{lo}{Colour fr the low LOR (mutual exclusive)}

\item{mid}{Colour for the mid LOR}

\item{hi}{Colour for the high LOR (co-occurence)}
}
\description{
Plots the Mutual exclusivity table directly from
cBioportal, or the matrix that was assembled using the sister
function
}
