% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visOneGene.R
\name{visOneGene}
\alias{visOneGene}
\title{visOneGene}
\usage{
visOneGene(gene, rge, mut, use.absolute = FALSE, purity = NULL,
  plot.cr = FALSE, use.seg = FALSE)
}
\arguments{
\item{mut}{}
}
\description{
visOneGene
}
