% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compTwoGenes.R
\name{compTwoGenes}
\alias{compTwoGenes}
\title{compTwoGenes}
\usage{
compTwoGenes(gene.x, gene.y, col.df, cex.type = "xy",
  plot.legend = TRUE)
}
\arguments{
\item{cex.type}{}
}
\description{
compTwoGenes
}
