% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lohLogistic.R
\name{plotLohProb}
\alias{plotLohProb}
\title{plotLohProb}
\usage{
plotLohProb(goi = NULL, soi = NULL, i = NULL, j = NULL,
  diff = FALSE, main = NULL, is.expr = FALSE, xlab = "Probability",
  minx = NULL)
}
\arguments{
\item{goi}{Gene of interest (e.g. ADAM10)}

\item{i}{if diff=TRUE; returned smoothed LOESS from plotLohProb()}

\item{j}{if diff=TRUE; returned smoothed LOESS from plotLohProb()}

\item{diff}{compares two LOESS segments}

\item{main}{Title}
}
\description{
plotLohProb
}
